Delete old download worker and rename newdownloadworker
